pyproject.toml
[tool.poetry]
name = "dikology-github-io"
version = "0.1.0"
description = ""
authors = ["dikology <dgimaletdinov@gmail.com>"]
readme = "README.md"
packages = [{include = "dikology"}]

[tool.poetry.dependencies]
python = "^3.10"
genanki = "^0.13.1"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
pandas = "^2.2.3"
numpy = "^2.2.0"
plotly = "^5.24.1"
psycopg2-binary = "^2.9.10"
jupysql = "^0.10.16"
toml = "^0.10.2"
nbformat = "^5.10.4"
duckdb = "^1.1.3"
argparse = "^1.4.0"
streamlit = "^1.41.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
